Add new events and style test files
